+++ aiida_castep/data/otfg.py
@@ -2,14 +2,44 @@
 Storing OTFG configuration as Data nodes
 """
 from __future__ import absolute_import
+from __future__ import print_function
 import re
-from aiida.orm import Data, GroupTypeString
+from aiida.orm import Data, Group, QueryBuilder
 from aiida.common.utils import classproperty
 from aiida.common import ValidationError
 from .utils import split_otfg_entry
+from .usp import UspData
 import six
 
-OTFGGROUP_TYPE = "data.castep.otfg.family"
+OLD_OTFGGROUP_TYPE = "data.castep.otfg.family"
+OTFGGROUP_TYPE = "castep.otfg"
+
+
+class OTFGGroup(Group):
+    """Class representing an OTFGGroup"""
+
+
+def migrate_otfg_family():
+    """Migrate the old OTFG families to new families"""
+    old_types = [OLD_OTFGGROUP_TYPE, "data.castep.usp.family"]
+    q = QueryBuilder()
+    q.append(Group, filters={'type_string': {'in': old_types}})
+
+    migrated = []
+    created = []
+    for (old_group, ) in q.iterall():
+        new_group, created = OTFGGroup.objects.get_or_create(
+            label=old_group.label, description=old_group.description)
+        new_group.add_nodes(list(old_group.nodes))
+        new_group.store()
+        migrated.append(new_group.label)
+        if created:
+            print("Created new style Group for <{}>".format(old_group.label))
+        else:
+            print(("Adding nodes to existing group <{}>".format(
+                old_group.label)))
+
+    return
 
 
 def upload_otfg_family(entries,
@@ -19,20 +49,16 @@ def upload_otfg_family(entries,
     """
     Set a family for the OTFG pseudo potential strings
     """
-    from aiida.orm import Group
     from aiida.common import UniquenessError, NotExistent
     from aiida.orm.querybuilder import QueryBuilder
     #from aiida.common import aiidalogger
 
     # Try to retrieve a group if it exists
     try:
-        group = Group.get(label=group_label, type_string=OTFGGROUP_TYPE)
+        group = OTFGGroup.get(label=group_label)
         group_created = False
     except NotExistent:
-        group = Group(
-            label=group_label,
-            type_string=OTFGGROUP_TYPE,
-        )
+        group = OTFGGroup(label=group_label, )
         group_created = True
 
     group.description = group_description
@@ -40,12 +66,14 @@ def upload_otfg_family(entries,
     otfg_and_created = []
     nentries = len(entries)
 
-    for s in entries:
+    for entry in entries:
         # Add it if it is just one existing data
-        if isinstance(s, OTFGData):
-            element, setting = s.element, s.string
-        else:
-            element, setting = split_otfg_entry(s)
+        if isinstance(entry, OTFGData):
+            element, setting = entry.element, entry.string
+        elif isinstance(entry, str):
+            element, setting = split_otfg_entry(entry)
+        elif isinstance(entry, UspData):
+            element, setting = entry.element, entry.md5sum
 
         qb = QueryBuilder()
         qb.append(OTFGData,
@@ -59,15 +87,33 @@ def upload_otfg_family(entries,
                   })
         existing_otfg = qb.first()
 
+        # Try find Usp data
         if existing_otfg is None:
 
-            if isinstance(s, OTFGData):
-                otfg_and_created.append((s, True))
-            else:
-                otfg, created = OTFGData.get_or_create(s,
+            qb = QueryBuilder()
+            qb.append(UspData,
+                      filters={
+                          'attributes.md5sum': {
+                              "==": setting
+                          },
+                          'attributes.element': {
+                              "==": element
+                          }
+                      })
+            existing_otfg = qb.first()
+
+        # Act based on wether the data exists
+        if existing_otfg is None:
+
+            if isinstance(entry, OTFGData):
+                otfg_and_created.append((entry, True))
+            elif isinstance(entry, str):
+                otfg, created = OTFGData.get_or_create(entry,
                                                        use_first=True,
                                                        store_otfg=False)
                 otfg_and_created.append((otfg, created))
+            elif isinstance(entry, UspData):
+                otfg_and_created.append((entry, True))
 
         else:
             if stop_if_existing:
@@ -83,7 +129,8 @@ def upload_otfg_family(entries,
     # Add other entries for the list to check
     if not group_created:
         for aiida_n in group.nodes:
-            if not isinstance(aiida_n, OTFGData):
+            if not isinstance(aiida_n, (OTFGData, UspData)):
+                print(("Warning: unsupported node: {}".format(aiida_n)))
                 continue
             elements.append((aiida_n.element, aiida_n.string))
 
@@ -96,12 +143,12 @@ def upload_otfg_family(entries,
         duplicates = set(
             [x for x in elements_names if elements_names.count(x) > 1])
         dup_string = ", ".join(duplicates)
-        raise UniquenessError("More than one OTFG found for the elements: " +
+        raise UniquenessError("More than one Nodes found for the elements: " +
                               dup_string + ".")
 
     # If we survive here uniqueness is fine
 
-    # Save the group
+    # Save the group - note we have not added the nodes yet
     if group_created:
         group.store()
 
@@ -110,12 +157,8 @@ def upload_otfg_family(entries,
     for otfg, created in otfg_and_created:
         if created:
             otfg.store()
-            #aiidalogger.debug("New node {} created for OTFG string {}".format(
-            #otfg.uuid, otfg.string))
         else:
             pass
-            #aiidalogger.debug("Reusing node {} for OTFG string {}".format(
-            #otfg.uuid, otfg.string))
 
     nodes_add = [otfg for otfg, created in otfg_and_created]
     nodes_new = [otfg for otfg, created in otfg_and_created if created is True]
@@ -258,10 +301,7 @@ class OTFGData(Data):
         """
         Return the OTFGData group with the given name.
         """
-        from aiida.orm import Group
-
-        return Group.objects.get(label=group_label,
-                                 type_string=cls.otfg_family_type_string)
+        return OTFGGroup.objects.get(label=group_label)
 
     @classmethod
     def get_otfg_groups(cls, filter_elements=None, user=None):
@@ -276,14 +316,13 @@ class OTFGData(Data):
                If defined, it should be either a DbUser instance, or a string
                for the user name (that is, the user email).
         """
-        from aiida.orm import Group
         from aiida.orm import QueryBuilder
         from aiida.orm import User
 
         query = QueryBuilder()
         filters = {'type_string': {'==': cls.otfg_family_type_string}}
 
-        query.append(Group, filters=filters, tag='group', project='*')
+        query.append(OTFGGroup, filters=filters, tag='group', project='*')
 
         if user:
             query.append(User,